feat(home): implement Home page and navigation with authentication handling
