+++ service/app/src/app/(home)/page.tsx
@@ -4,16 +4,16 @@ import { ErrorBoundary } from "react-error-boundary"
 
 import { useAuthStore } from "@/entities/auth"
 
-import { GameSection } from "../widgets/GameSection"
-import { HeroSection } from "../widgets/HeroSection"
-import { HomeNavigation } from "../widgets/HomeNavigation"
-import ErrorPage from "./error"
+import { GameSection } from "../../widgets/GameSection"
+import { HeroSection } from "../../widgets/HeroSection"
+import { HomeNavigation } from "../../widgets/HomeNavigation"
+import ErrorPage from "../error"
 
 function HomeContent() {
   const { isAuthenticated } = useAuthStore()
 
   return (
-    <main className="flex min-h-screen w-full flex-col items-start gap-112 bg-background-primary p-0">
+    <main className="flex min-h-screen w-full flex-col items-start gap-112 bg-background-primary">
       <HomeNavigation isLoggedIn={isAuthenticated} />
       <HeroSection />
       <GameSection />
